Update patch set 1
Patch Set 1:

I don't think this is related. The -exact option just means that we don't discard RGB values when A = 0. The compression quality of the alpha is not taken into account (maybe it will save a bit more RGB values because some pixels are no longer fully transparent because of lower alpha quality but that's all).

Patch-set: 1
CC: Gerrit User 1289110 <1289110@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>