Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > hmm... the agreement was that we would rescale nevertheless even if the dimension were the same to provide a continuous experience without hiccup when dimension are changed (and also if someone wants to slightly blur the picture by rescaling it to its exact dimension)
> 
> I thought that might have been the case, just wanted to double check. shall we call crbug.com/webp/397 WAI?

I don't think so. The issue with this bug is that resizing discards RGB values for fully transparent pixels, even if -exact is used (which just means that we keep all color data even for invisible areas). I'm working on a change to disable pre-multiplication of RGB by A during a resize if -exact is passed.

Patch-set: 1
CC: Gerrit User 1289110 <1289110@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>