Update patch set 6
Patch Set 6:

Maybe we should take a step back on this one before going further. I think for 1.0.1 this is a fairly large change and though the intent is fairly clear, there's a chance we'll miss something. Since it's always been broken it would be safest to let this bake a bit for a 1.0.2. As for the change I wonder if we could avoid the mass GCHAR comparisons and just convert the file parameters as needed.

Patch-set: 6
Reviewer: Gerrit User 1002219 <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>