Update patch set 8
Patch Set 8:

I tried the typical -msse* clang flags and they are not supported by clang-cl. The /arch are.

The SIMD availability is tested twice: first with no flag, and if that does not work, with flags. With no flag, MSVC detects SSE4 (meaning by default, it detects it). BUT, clang-cl does not .....

So, it seems that we have to keep the flags for now.

Patch-set: 8
Reviewer: Gerrit User 1140964 <1140964@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>