src: alpha flag in the VP8X chunk is not mandatory.
The alpha flag of the VP8X chunk seems mostly informational. In
particular even when an image has an alpha channel while no alpha flag,
it can still be perfectly loaded. Therefore this check is not proper.

This is even more true as libwebp does not fill the VP8X chunk when
exporting in lossless. So we end up with images exported by libwebp
which fail to be loaded by the same library!

Change-Id: Ib02453a94563865492cd09a187fb12efc6919fcb