Update patch set 1
Patch Set 1:

Adding some background comment: in GIMP, we had a bug report about a file created by libwebp as lossless, which fails to be opened (also through libwebp).

https://gitlab.gnome.org/GNOME/gimp/issues/1818

So looking at the code, I realize that there are 2 very different code path for encoding WebP images whether they are lossless or not. Cf. `WebPEncode()` code: when lossy, it calls `VP8EncWrite()`, which in turns calls `PutWebPHeaders()` and sets the VP8X chunk.

For lossless export, `VP8LEncodeImage()` doesn't do this.

Now I noticed this text in the docs: "Background color MAY contain a transparency value (alpha), even if the Alpha flag in VP8X chunk is unset."
https://developers.google.com/speed/webp/docs/riff_container#extended_file_format

I also confirm that just deleting this check (i.e. accepting an image with alpha, yet an unset alpha flag in the VP8X chunk) works great, i.e. that the image loads fine in GIMP just as it was at creation time.

Note 1: maybe it would be worth also appropriately setting the alpha flag (as any other flag in the VP8X header) when creating lossless images too. This could be a further change. Even though, refusing an image without this flag doesn't make sense since even your docs says it is optional and the image loads fine otherwise.

Note 2: the bug report on our tracker also says the image loads with artifacts in Chrome. I want to note that this is also most likely a different bug (in Chrome). Indeed once we remove the check and loads the image in GIMP, it is perfectly fine (no artifacts whatsoever). So the bug is not in libwebp (at least not in master code as I assume Chrome uses libwebp).

Patch-set: 1
Reviewer: Gerrit User 1175869 <1175869@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>