Warn when alpha flag missing in VP8X header on image with alpha.
It would seem that such images should be considered as invalid. Yet they
can still be opened properly, and we know for a fact that such images
are already in nature as libwebp was producing these (apparently because
of a bug in Exiv2). So let's still accept these, and simply produce a
warning.

Change-Id: I7ee81894a79ee2eaccc87bc2ca87a3945b6b3375