Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 2:
> > 
> > Thanks for updating the patch. I'll upload one more to land it.
> > Note it can be helpful to reply to the comments in the change when pushing a new patch so reviewers know it's safe to take another look.
> 
> Thank you. Apologies for the lack of replies – I hadn't figured out the appropriate way to do that.

No problem, it's pretty common to get lost in the interface. Each comment will have a 'reply' and 'done' links that you can use for that (some screenshots from the old ui [1]). Those are only stored as drafts though so you have to make sure to hit the reply button at the top of the page to publish them.

[1] https://gerrit-review.googlesource.com/Documentation/user-review-ui.html#inline-comments

Patch-set: 4