Update patch set 1
Patch Set 1:

Right, your patch is because we set LOCAL_GCC_VERSION to 0 by default (when __GNUC__ is not defined), which triggers this condition.
That's the only spot where we use a <= for LOCAL_GCC_VERSION (otherwise, it's a ==).
So, LGTM. I'll wait for others to comment.

Patch-set: 1
Reviewer: Gerrit User 1140964 <1140964@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>