anim_diff: normalize `ok = ... && ok` statements
`ok &= ...` is more common in the codebase than `ok = ... && ok` when
accumulating a result while unconditionally executing functions. (`ok =
ok &&` is used in cases that should short circuit.) In this case
multiple checks may fail and their error messages may aid in debugging.

This will also improve the formatting when clang-format is applied to
the codebase.

Bug: 433996651
Change-Id: Ie4e2908b857122d90f6e93f06b10cb48dc86b18e